fix add/update room service, hover to show tooltip

+++ admin/src/pages/bookingorder/BookingOrder.jsx
@@ -1,13 +1,13 @@
 import { React, useContext, useState, useEffect } from "react";
 import { AuthContext } from "../../contexts/AuthContext";
-import { Container, Row, Table, Button } from 'reactstrap';
+import { Container, Row, Table, Button, UncontrolledTooltip } from 'reactstrap';
 import useFetch from "../../hooks/useFetch";
 import instance from "../../service";
 import { faTrashCan } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 const BookingOrder = () => {
     const { user, dispatch } = useContext(AuthContext);
-    const  {data,reFetch}= useFetch("/booking/")
+    const { data, reFetch } = useFetch("/booking/")
     const [dataBookingOrder, setDataBookingOrder] = useState([])
     useEffect(() => {
         setDataBookingOrder(data);
@@ -15,14 +15,19 @@ const BookingOrder = () => {
         [data]);
     const handleConfirm = async (selectedBookingOrder) => {
         // console.log("selected booking:",selectedBookingOrder)
-        await instance.put("/booking/update/"+selectedBookingOrder._id,{bookingStatus:2});
+        await instance.put("/booking/update/" + selectedBookingOrder._id, { bookingStatus: 2 });
         reFetch();
     }
     const handleDelete = async (selectedBookingOrder) => {
-        // console.log("selected booking:",selectedBookingOrder)
-        await instance.delete("/booking/delete/"+selectedBookingOrder._id);
-        reFetch();
-    } 
+        if (window.confirm('Do you want to delete this order?') === true) {
+            try {
+                await instance.delete("/booking/delete/" + selectedBookingOrder._id);
+                reFetch();
+            } catch (error) {
+            }
+        } else {}
+      
+    }
     return (
         <div className="BookingOrder">
             {user ?
@@ -55,11 +60,17 @@ const BookingOrder = () => {
                                                     <td><span style={{ color: "green" }}>Success</span></td> : null
                                                 }
                                                 {booking.bookingStatus === 1 ?
-                                                    <td><Button color="warning" onClick={()=>handleConfirm(booking)}>Confirm</Button></td> : null
+                                                    <td><Button color="warning" onClick={() => handleConfirm(booking)}>Confirm</Button></td> : null
                                                 }
                                                 {booking.bookingStatus === 2 ?
-                                                    <td><span style={{ color: "red" }}>Cancel --{'>'}<FontAwesomeIcon onClick={()=>handleDelete(booking)} icon={faTrashCan}/></span></td> : null
+                                                    <td><span style={{ color: "red" }}>Cancel <FontAwesomeIcon id="UncontrolledTooltipExample" onClick={() => handleDelete(booking)} icon={faTrashCan} /></span></td> : null
                                                 }
+                                                <UncontrolledTooltip
+                                                    placement="top"
+                                                    target="UncontrolledTooltipExample"
+                                                >
+                                                    Delete this order?
+                                                </UncontrolledTooltip>
                                             </>
                                         </tr>
